G2/app/ai/xray_analysis.py
"""
X-Ray Analysis Module

Analyzes chest X-ray images and provides diagnostic suggestions.
Currently uses mock implementation for development.
In production, integrate with trained medical imaging models.
"""

from typing import Dict
import logging

logger = logging.getLogger(__name__)


def analyze_xray(image_bytes: bytes) -> Dict[str, any]:
    """
    Analyze X-ray image and provide diagnostic suggestion.
    
    This is a MOCK implementation for development/testing.
    In production, replace with actual computer vision model.
    
    Args:
        image_bytes: X-ray image file content in bytes
        
    Returns:
        dict: Analysis result with keys:
            - diagnosis: str - Diagnostic finding
            - confidence: float - Confidence score (0.0 to 1.0)
            - findings: List[str] - Specific findings
            - notes: str - Additional observations
            
    Raises:
        ValueError: If image_bytes is empty or invalid
        Exception: If analysis fails
        
    Production Integration Example:
        ```python
        import torch
        from PIL import Image
        from io import BytesIO
        from torchvision import transforms
        
        # Load pre-trained model (e.g., CheXNet, DenseNet)
        model = torch.load('xray_model.pth')
        model.eval()
        
        # Preprocess image
        image = Image.open(BytesIO(image_bytes)).convert('RGB')
        transform = transforms.Compose([
            transforms.Resize(224),
            transforms.CenterCrop(224),
            transforms.ToTensor(),
            transforms.Normalize([0.485, 0.456, 0.406], [0.229, 0.224, 0.225])
        ])
        input_tensor = transform(image).unsqueeze(0)
        
        # Run inference
        with torch.no_grad():
            output = model(input_tensor)
            probabilities = torch.softmax(output, dim=1)
            
        # Map to diagnoses
        class_names = ['Normal', 'Pneumonia', 'COVID-19', 'Tuberculosis', ...]
        confidence, predicted_class = torch.max(probabilities, 1)
        
        return {
            "diagnosis": class_names[predicted_class.item()],
            "confidence": confidence.item(),
            "findings": [...],
            "notes": "..."
        }
        ```
    """
    if not image_bytes:
        raise ValueError("Image bytes cannot be empty")
    
    if len(image_bytes) < 1000:
        raise ValueError("Image file too small, possibly corrupted")
    
    logger.info(f"Analyzing X-ray image ({len(image_bytes)} bytes)")
    
    # Validate basic image format
    if not _validate_image_format(image_bytes):
        raise ValueError("Invalid image format. Expected JPEG or PNG.")
    
    # MOCK IMPLEMENTATION
    # Returns realistic diagnostic suggestions for testing
    
    # Simulate different findings based on image size (for variety in testing)
    image_hash = len(image_bytes) % 5
    
    mock_results = [
        {
            "diagnosis": "Clear chest X-ray with no acute findings",
            "confidence": 0.92,
            "findings": [
                "Lungs are clear bilaterally",
                "No pleural effusion",
                "Cardiac silhouette normal size",
                "No pneumothorax visible"
            ],
            "notes": "Normal chest radiograph. No acute cardiopulmonary abnormality detected."
        },
        {
            "diagnosis": "Possible pneumonia in right lower lobe",
            "confidence": 0.78,
            "findings": [
                "Increased opacity in right lower lung field",
                "Possible air bronchograms",
                "Costophrenic angles clear",
                "No pleural effusion"
            ],
            "notes": "Findings suggestive of right lower lobe pneumonia. Clinical correlation recommended. Consider follow-up imaging after treatment."
        },
        {
            "diagnosis": "Mild cardiomegaly",
            "confidence": 0.85,
            "findings": [
                "Enlarged cardiac silhouette",
                "Cardiothoracic ratio >0.5",
                "Lungs clear",
                "No pulmonary congestion"
            ],
            "notes": "Cardiac enlargement noted. Lungs appear clear. Consider echocardiogram for further cardiac assessment."
        },
        {
            "diagnosis": "Small pleural effusion on left side",
            "confidence": 0.81,
            "findings": [
                "Blunting of left costophrenic angle",
                "Small amount of pleural fluid",
                "Right lung clear",
                "No consolidation"
            ],
            "notes": "Small left pleural effusion. Clinical correlation needed. May require ultrasound for further characterization."
        },
        {
            "diagnosis": "Chronic changes consistent with COPD",
            "confidence": 0.74,
            "findings": [
                "Hyperinflation of lungs",
                "Flattened diaphragms",
                "Increased retrosternal airspace",
                "No acute infiltrates"
            ],
            "notes": "Findings suggestive of chronic obstructive pulmonary disease. No acute exacerbation visible. Clinical correlation advised."
        }
    ]
    
    result = mock_results[image_hash]
    
    logger.info(f"X-ray analysis completed: {result['diagnosis']} (confidence: {result['confidence']:.2f})")
    return result


def _validate_image_format(image_bytes: bytes) -> bool:
    """
    Validate image file format.
    
    Checks for JPEG or PNG format using magic numbers.
    
    Args:
        image_bytes: Image file content
        
    Returns:
        bool: True if valid JPEG or PNG
    """
    if not image_bytes or len(image_bytes) < 10:
        return False
    
    # Check JPEG magic number (FF D8 FF)
    if image_bytes[:3] == b'\xff\xd8\xff':
        return True
    
    # Check PNG magic number (89 50 4E 47 0D 0A 1A 0A)
    if image_bytes[:8] == b'\x89PNG\r\n\x1a\n':
        return True
    
    return False


def get_supported_findings() -> list:
    """
    Get list of findings the model can detect.
    
    Returns:
        list: Supported diagnostic findings
    """
    return [
        "Normal",
        "Pneumonia",
        "Pleural effusion",
        "Cardiomegaly",
        "Pneumothorax",
        "Mass/Nodule",
        "COPD/Emphysema",
        "Atelectasis",
        "Pulmonary edema",
        "Consolidation"
    ]
