G2/requirements.txt
alembic==1.13.1
annotated-types==0.7.0
anyio==4.12.1
bcrypt==4.0.1
black==24.1.1
certifi==2026.1.4
cffi==2.0.0
click==8.3.1
colorama==0.4.6
cryptography==42.0.0
dnspython==2.8.0
ecdsa==0.19.1
email-validator==2.1.0
fastapi==0.109.0
flake8==7.0.0
greenlet==3.3.1
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.26.0
idna==3.11
iniconfig==2.3.0
Mako==1.3.10
MarkupSafe==3.0.3
mccabe==0.7.0
mypy==1.8.0
mypy_extensions==1.1.0
packaging==26.0
passlib==1.7.4
pathspec==1.0.4
pillow==12.1.1
platformdirs==4.5.1
pluggy==1.6.0
psycopg==3.3.2
psycopg-binary==3.3.2
psycopg2-binary==2.9.10
pyasn1==0.6.2
pycodestyle==2.11.1
pycparser==3.0
pydantic==2.12.5
pydantic-settings==2.1.0
pydantic_core==2.41.5
pyflakes==3.2.0
PyJWT==2.8.0
PyPDF2==3.0.1
pytest==7.4.4
pytest-asyncio==0.23.3
python-dateutil==2.8.2
python-dotenv==1.2.1
python-jose==3.3.0
python-magic==0.4.27
python-multipart==0.0.6
pytz==2024.1
PyYAML==6.0.3
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.46
starlette==0.35.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
uvicorn==0.27.0
watchfiles==1.1.1
websockets==16.0
